# Optimize the rowwise add function.